refactor(api): Restructure dashboard API and add operations endpoints
- Restructured API to improve code organization and maintainability.
- Introduced new API endpoints for managing and monitoring operations.
- Moved account management, authentication, API key, avatar, and OTP related endpoints to separate files for better modularity.
